pyproject.toml
[project]
name = "fastapi-project"
version = "1.0.0"
description = "FastAPI project with JWT Auth, MySQL, WebSocket, and Celery"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["fastapi", "jwt", "websocket", "celery", "mysql"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.1",
    "websockets==14.1",
    "sqlalchemy==2.0.36",
    "pymysql==1.1.1",
    "python-jose[cryptography]==3.3.0",
    "bcrypt==4.2.1",
    "python-multipart==0.0.20",
    "python-dotenv==1.0.1",
    "alembic==1.14.0",
    "celery==5.4.0",
    "redis==5.2.1",
    "pydantic[email]==2.10.3",
    "pydantic-settings==2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "httpx==0.28.1",
    "pytest-cov==6.0.0",
    "faker==33.1.0",
    "ruff==0.8.4",
    "mypy==1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["F401", "F811"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=70",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["app"]